🚚 move database implementation of projection repository in event-listeners crate
