+++ event-listeners/Cargo.toml
@@ -16,6 +16,14 @@ async-trait = "0.1.56"
 futures = {version = "0.3.21", features = ["alloc"]}
 tokio = {version = "1.18.2", features = ["full"]}
 
+# Database
+diesel = {version = "1.4.8", default-features = false, features = [
+  "postgres",
+  "serde_json",
+  "uuidv07",
+  "chrono",
+]}
+
 # Core types
 uuid = {version = "0.8.2", default_features = false, features = [
   "v4",
