+++ event-listeners/src/infrastructure/database/models/mod.rs
@@ -1,0 +1,3 @@
+#[allow(clippy::extra_unused_lifetimes)]
+mod contributors;
+pub use contributors::*;
