+++ event-listeners/src/infrastructure/database/repositories/mod.rs
@@ -1,0 +1,1 @@
+mod contributor;
