+++ event-listeners/src/infrastructure/mod.rs
@@ -1,1 +1,2 @@
+pub mod database;
 pub mod github;
