+++ marketplace-infrastructure/src/database/models/mod.rs
@@ -1,7 +1,3 @@
 #[allow(clippy::extra_unused_lifetimes)]
-mod contributors;
-pub use contributors::*;
-
-#[allow(clippy::extra_unused_lifetimes)]
 mod payment_request;
 pub use payment_request::*;
