+++ marketplace-infrastructure/src/database/repositories/mod.rs
@@ -1,2 +1,1 @@
-mod contributor;
 mod payment_request;
