--- marketplace-infrastructure/src/database/tests/mod.rs
@@ -1,1 +1,0 @@
-mod contributor_repository;
